API

Enlsip.AbstractCnlsModel — Type

AbstractCnlsModel

Abstract type for CnlsModel structure.

Enlsip.CnlsModel — Type

model = CnlsModel(residuals, nb_parameters, nb_residuals)

Constructor for CnlsModel

The following arguments are mandatory to instantiate a model:

* `residuals` : function that computes the vector of residuals

* `nb_parameters` : number of variables

* `nb_residuals` : number of residuals

The following can be provided as optionnal arguments:

* `stating_point` : initial solution (default is a vector of zeros of appropriate dimension)

* `jacobian_residuals` : function that computes the jacobian matrix of the residuals. If not passed as argument, it is computed numericcaly by forward differences

* `eq_constraints` : function that computes the vector of equality constraints

* `jacobian_eqcons` : function that computes the jacobian matrix of the equality constraints. If not passed as argument, it is computed numericcaly by forward differences

* `nb_eqcons` : number of equality constraints

* `ineq_constraints` : function that computes the vector of inequality constraints

* `jacobian_ineqcons` : function that computes the jacobian matrix of the inequality constraints. If not passed as argument, it is computed numericcaly by forward differences

* `nb_ineqcons` : number of inequality constraints

* `x_low` and `x_upp` : respectively vectors of lower and upper bounds

Enlsip.CnlsModel — Type

CnlsModel

Structure modeling an instance of a constrainted nonlinear least squares problem.

* `residuals` : function that computes the vector of residuals

* `nb_parameters` : number of variables

* `nb_residuals` : number of residuals

* `stating_point` : initial solution

* `jacobian_residuals` : function that computes the jacobian matrix of the residuals

* `eq_constraints` : function that computes the vector of equality constraints

* `jacobian_eqcons` : function that computes the jacobian matrix of the equality constraints

* `nb_eqcons` : number of equality constraints

* `ineq_constraints` : function that computes the vector of inequality constraints

* `jacobian_ineqcons` : function that computes the jacobian matrix of the inequality constraints

* `nb_ineqcons` : number of inequality constraints

* `x_low` and `x_upp` : respectively vectors of lower and upper bounds

* `status_code` : integer indicating the solving status of the model.

Enlsip.objective_value — Method

objective_value(model)

Once the given model has been solved, returns the value of the objective function, i.e. sum of squared residuals functions, computed at the optimal solution. If no convergence, this value is computed at the last solution obtained.

See also: CnlsModel and solution.

Enlsip.solution — Method

solution(model)

Once the given model has been solved, this function returns the optimal solution, or last solution obtained if no convergence, as a Vector of approriate dimension.

See also: solve! and CnlsModel.

Enlsip.solve! — Method

solve!(model)

Once a CnlsModel has been instantiated, this function solves the optimzation problem associated by using the method implemented in Enlsip.

The following optionnal arguments can be provided:

silent::Bool

Set to true if one wants the algorithm to print details about the iterations and termination of the solver

Default value is set to false

max_iter::Int

Maximum number of iterations allowed

Default value is set to 100

scaling::Bool

Set to true if one wants the algorithm to work with a constraints jacobian matrix whose rows are scaled (i.e. all constraints gradients vectors are scaled)

Default value is set to false

Enlsip.status — Method

status(model)

Returns readable information on the solving status of model.

See also:CnlsModel.

Enlsip.total_nb_constraints — Method

total_nb_constraints(model)

Returns the total number of constraints, i.e. equalities, inequalities and bounds, of the given model.

See also: CnlsModel.