Usage

This sections provides details on how to instantiate and solve a constrained least squares problem with Enlsip.jl As a reminder from Home, problems to solve are of the following form:

\[\begin{aligned} \min_{x \in \mathbb{R}^n} \quad & \dfrac{1}{2} \|r(x)\|^2 \\ \text{s.t.} \quad & c_i(x) = 0, \quad i \in \mathcal{E} \\ & c_i(x) \geq 0, \quad i \in \mathcal{I}, \\ \end{aligned}\]

with:

the residuals $r_i:\mathbb{R}^n\rightarrow\mathbb{R}$ and the constraints $c_i:\mathbb{R}^n\rightarrow\mathbb{R}$ assumed to be $\mathcal{C}^2$ functions;

norm $\|\cdot\|$ denoting the Euclidean norm.

Note that with this formulation, bounds constraints are not distinguished from general inequality constraints. Though, as shown later in this section, they can be provided as vectors of lower and/or upper bounds, which is more convenient for this type of constraints.

It should be borne in mind, however, that the method implemented in Enlsip.jl has been conceived for nonlinear problems, as there is no other assumption made about the nature of the residuals and constraints functions, apart from being two-time continously differentiable. The algorithm can still be used to solve linear least squares subject to linear constraints but it will not be as effective as other software where those aspects are taken into account in the design of the optimization method.

Instantiate a model

Solving a problem with Enlsip is organized in two steps.

First, a model of type CnlsModel must be instantiated.

The CnlsModel constructor requires the evaluation functions of residuals, constraints, their associated jacobian matrices and dimensions of the problem.

Although the package enables one to create linear unconstrained least squares, it is recommended to use it to solve nonlinear least squares with general constraints.

The three following positional arguments are mandatory to create a model:

residuals : function that computes the vector of residuals

nb_parameters : number of variables

nb_residuals : number of residuals

The following keywords arguments are optional and deal with constraints and Jacobian matrices. If the Jacobian matrices functions are not provided, they are computed numerically by forward differences within the Enlsip solver.

Argument	Details

starting_point	initial solution (can be an infeasbile point)

jacobian_residuals	function computing the Jacobian matrix of the residuals

eq_constraints	function computing the equality constraints

jacobian_eqcons	function computing the Jacobian matrix of the equality constraints

nb_eqcons	number of equality constraints

ineq_constraints	function computing the inequality constraints

jacobian_ineqcons	function computing the Jacobian matrix of the inequality constraints

nb_ineqcons	number of inequality constraints

x_low	vector of lower bounds

x_upp	vector of upper bounds

Solving a model

Then, the Enlsip solver can be used by calling the solve! function on a instantiated model.

Enlsip.solve! — Function

solve!(model)

Once a CnlsModel has been instantiated, this function solves the optimzation problem associated by using the method implemented in the Enlsip solver.

The following optionnal arguments can be provided:

silent::Bool

Set to false if one wants the algorithm to print details about the iterations and termination of the solver

Default value is set to false

max_iter::Int

Maximum number of iterations allowed

Default value is set to 100

scaling::Bool

Set to true if one wants the algorithm to work with a constraints jacobian matrix whose rows are scaled (i.e. all constraints gradients vectors are scaled)

Default value is set to false

By default, the algorithm will print some details about the iterations performed through during execution.

Here are some details on how to read and understand the different columns of the output:

Column	Details

iter	iteration number

objective	value of the sum of squared residuals (i.e. objective function) at current point

$\vert\vert$ active_constraints $\vert\vert^2$	value of the sum of squared active constraints at current point

$\vert\vert$ p $\vert\vert$	norm of the search direction computed at current iteration

$\alpha$	value of the steplength computed at current iteration

reduction	reduction in the objective function performed at curret iteration

Once the Enlsip solver has been called, one can get additional info about the success, or failure, of the algorithm by calling one of the following functions:

Function

solution

status

objective_value

Enlsip.solution — Function

solution(model)

Once the given model has been solved, this function returns the optimal solution, or last solution obtained if no convergence, as a Vector of approriate dimension.

Enlsip.status — Function

status(model)

This functions returns a Symbol that gives brief information on the solving status of model.

If a model has been instantiated but the solver has not been called yet, it will return :unsolved.

Once the solver has been called and if a first order critical point satisfying the convergence criteria has been computed, it will return :successfully_solved.

If the algorithm met an abnornakl termination criteria, it will return one of the following:

:failed : the algorithm encoutered a numerical error that triggered termination

:maximum_iterations_exceeded : a solution could not be reached within the maximum number of iterations.

Enlsip.objective_value — Function

objective_value(model)

Once the given model has been solved, returns the value of the objective function, i.e. sum of squared residuals functions, computed at the optimal solution. If no convergence, this value is computed at the last solution obtained.

Examples

Problem 65 from Hock and Schittkowski collection[1]

We show how to implement and solve the following problem:

\[\begin{aligned} \min \quad & (x_1-x_2)^2 + \dfrac{(x_1+x_2-10)^2}{9}+(x_3-5)^2 \\ \text{s.t.} \quad & 48-x_1^2-x_2^2-x_3^2 \geq 0\\ & -4.5\leq x_i \leq 4.5, \quad i=1,2\\ & -5 \leq x_3 \leq 5. \end{aligned}.\]

First, we provide the dimensions of the problems.

# Dimensions of the problem

n = 3 # number of parameters
m = 3 # number of residuals
nb_eq = 0 # number of equality constraints
nb_constraints = 7 # number of inequality constraints

Then, we define the functions required to compute the residuals, constraints, their respective jacobian matrices and a starting point.

# Residuals and Jacobian matrix associated
r(x::Vector) = [x[1] - x[2]; (x[1]+x[2]-10.0) / 3.0; x[3]-5.0]

jac_r(x::Vector) = [1. -1. 0;
    1/3 1/3 0.;
    0. 0. 1.]

# Constraints (one equality and box constraints)

c(x::Vector) = [48.0 - x[1]^2-x[2]^2-x[3]^2] # evaluation function for the equality constraint
jac_c(x::Vector) = [ -2x[1] -2x[2] -2x[3]] # Jacobian matrix of the equality constraint

x_l = [-4.5, -4.5, -5.0] # lower bounds
x_u = [4.5, 4.5, 5.0] # upper bounds

# Starting point
x0 = [-5.0, 5.0, 0.0]

A CnlsModel can now be instantiated.

# Instantiate a model associated with the problem
hs65_model = Enlsip.CnlsModel(r, n, m ;starting_point=x0, ineq_constraints = c,
nb_ineqcons = 1, x_low=x_l, x_upp=x_u, jacobian_residuals=jac_r, jacobian_ineqcons=jac_c)

Finally, the solve! function can be called on our model. In this example, we set the silent optional argument to true so no output will be printed. See previous Solving a model section for details about what is printed after solving a problem. Other optional arguments remain to default values.

Enlsip.solve!(hs65_model, silent=true)

Once Enlsip solver has been executed on a problem, one can check if the problem has been successfully solved or not.

Enlsip.status(hs65_model)

:successfully_solved

One can also get the optimal solution obtained and value of objective function at that point.

hs65_solution = Enlsip.solution(hs65_model)

3-element Vector{Float64}:
 3.650461726852526
 3.6504617268525554
 4.620417552729356

hs65_objective = Enlsip.objective_value(hs65_model)

0.9535288568047827

1W. Hock and K. Schittkowski. Test Examples for Nonlinear Programming Codes, volume 187 of Lecture Notes in Economics and Mathematical Systems. Springer, second edition, 1980.